Update sshell.c
Decided to pass a Cmd struct to sys function instead of char array.
ie, pulled out the creation of Cmd struct from this sys() function to main.
We decided this would be ideal as we would need to use the Struct in other areas from main.
Professor also mentioned to divide code into phases of (1) parsing and (2) executing